refresh no longer stops darkmode
